hash

+++ include/flame/universe/entity.h
@@ -23,7 +23,7 @@ namespace flame
 
 		virtual Entity* get_parent() const = 0;
 
-		virtual Component* get_component(uint hash) const = 0;
+		virtual Component* get_component(uint64 hash) const = 0;
 		virtual void add_component(Component* c) = 0;
 		virtual void remove_component(Component* c, bool destroy = true) = 0;
 		virtual void remove_all_components(bool destroy = true) = 0;
